Move the apply command to a function * move the redundant command to a new function * move the dryrun files to dryrun package
